n2mobil_backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
drf-yasg==1.21.5
psycopg2-binary==2.9.6
django-redis==5.2.0
coreapi==2.3.3
coreschema==0.0.4
PyYAML==6.0
markdown==3.4.1
pytz==2023.3
asgiref==3.6.0
sqlparse==0.4.4
redis==5.0.0
python-decouple>=3.8
djangorestframework-api-key
